+++ src/compdd/dock6/dock6_pipeline.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from compdd.configs.docking_config import RootConfig
+from compdd.configs.root_config import RootConfig
 from compdd.ligands.ligands_prep import ligands_prep
 from compdd.dock6._dock6_prep_rec import dock6_prep_rec
 from compdd.dock6._dock6_docking import dock6_docking
@@ -13,6 +13,8 @@ class DOCK6Pipeline():
     cfg: RootConfig
 
     def run(self):
+        self.cfg.common.program = "dock6"
+
         lig_files = ligands_prep(self.cfg)
         prepped_recs = dock6_prep_rec(self.cfg)
         pairs = matchmixer(prepped_recs, lig_files, self.cfg.common.prepared_suffix)